Fixed DB corrpution in client. Minor change in pool4
There was a problem with the db while using client_for_each because it
got corrupted, it changed the db. Fixed that problem by putting the
original information within a dummy variable and just before the cycle
ended, made the value of the original address the one of the dummy(the
original).
Changed breaks to returns so the loop ends and prevent useless
iterations.